refactor: create a constants.js module and move relevant things inside
