+++ src/constants.js
@@ -1,0 +1,19 @@
+/* Constants */
+
+export const SCALE = 4;
+export const SCALE2 = 2;
+export const SPRITE_WIDTH = 8;
+export const GRID_WIDTH_PX = SCALE * SCALE2 * SPRITE_WIDTH;
+
+export const FPS = 60;
+
+export const VALID_CONTROLLER_KEYS = [
+	"ArrowLeft",
+	"ArrowRight",
+	"ArrowUp",
+	"ArrowDown",
+	"KeyZ",
+	"KeyX",
+];
+export const SCREEN_WIDTH_PX = 960;
+export const SCREEN_HEIGHT_PX = 540;
