+++ src/data.js
@@ -1,9 +1,9 @@
 /* Data Module */
 
-export const SCALE = 4;
-export const SCALE2 = 2;
-export const SPRITE_WIDTH = 8;
-export const GRID_WIDTH_PX = SCALE * SCALE2 * SPRITE_WIDTH;
+// export const SCALE = 4;
+// export const SCALE2 = 2;
+// export const SPRITE_WIDTH = 8;
+// export const GRID_WIDTH_PX = SCALE * SCALE2 * SPRITE_WIDTH;
 
 const TEST_LEVEL = [
 	[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
